requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
cryptography==42.0.5
dj-rest-auth==5.0.2
Django==5.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-geoposition==0.3.0
djangorestframework==3.14.0
drf-extensions==0.7.1
drf-standardized-errors==0.13.0
gcloud==0.18.3
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauth2client==4.1.3
pillow==10.2.0
protobuf==4.25.3
psycopg==3.1.18
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.2
python-decouple==3.8
python-jwt==4.1.0
pytz==2024.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.1
